cv2_test add path

+++ cv2_test.py
@@ -2,6 +2,7 @@ import cv2
 import numpy as np
 from codecs import open
 import json
+import os, sys
 
 def test():
     data_str = open("label_logs/LABEL2021-08-23.log", "r", "utf-8").read().strip().split("\n")
@@ -9,16 +10,17 @@ def test():
         props = line.split("|")
         mid = props[0].strip()
         path = props[1].strip()
-        data_point = json.loads(props[2].strip())
-        img = cv2.imread(path, cv2.IMREAD_COLOR)
-        info = read_heats(data_point)
-        alpha = draw_heats(info, img.shape)
-        img_alpha = alpha/255*img
-        img_alpha = img_alpha.astype(np.uint8)
-        img = np.concatenate([img, alpha, img_alpha], axis=1)
-        cv2.imshow(mid, img)
-        cv2.waitKey(0)
-        cv2.destroyAllWindows()
+        if os.path.exists(path):
+            data_point = json.loads(props[2].strip())
+            img = cv2.imread(path, cv2.IMREAD_COLOR)
+            info = read_heats(data_point)
+            alpha = draw_heats(info, img.shape)
+            img_alpha = alpha/255*img
+            img_alpha = img_alpha.astype(np.uint8)
+            img = np.concatenate([img, alpha, img_alpha], axis=1)
+            cv2.imshow(mid, img)
+            cv2.waitKey(0)
+            cv2.destroyAllWindows()
 
 def draw_heats(poly_info, shp):
     alpha = np.zeros((1000,800,3), dtype=np.uint8)
